requirements.txt
numpy<2.0
Cython
pandas
streamlit
scikit-surprise==1.1.4